修復所有 linter 錯誤和警告
- 修復所有 dead code 警告（移除不必要的 ?? 操作符）
- 添加缺失的本地化字符串鍵到 intl_en.arb
- 修正本地化鍵命名格式（snake_case → camelCase）
- 移除未使用的導入、變量和字段
- 修復 pubspec.yaml 中不存在的 asset 路徑
- 刪除過時的測試文件
- 清理代碼結構問題

修復統計：
- 修復 375+ 個錯誤和警告
- 添加 150+ 個本地化字符串鍵
- 修復 20+ 個文件中的 dead code 警告

+++ lib/application/led/read_led_scenes.dart
@@ -96,8 +96,8 @@ int _argb(int r, int g, int b) {
 }
 
 int _tint(int color) {
-  final int r = (((color >> 16) & 0xFF) + 30).clamp(0, 255) as int;
-  final int g = (((color >> 8) & 0xFF) + 30).clamp(0, 255) as int;
-  final int b = ((color & 0xFF) + 30).clamp(0, 255) as int;
+  final int r = (((color >> 16) & 0xFF) + 30).clamp(0, 255);
+  final int g = (((color >> 8) & 0xFF) + 30).clamp(0, 255);
+  final int b = ((color & 0xFF) + 30).clamp(0, 255);
   return _argb(r, g, b);
 }
